# apps/api/src/modules/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: BirthPad API
  version: 0.2.0
servers:
  - url: http://localhost:3000
paths:
...
  /config/alerts/history:
    get:
      summary: Get persisted alert action history
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Alert action history items
  /config/alerts/{id}/ack:
    post:
      summary: Acknowledge an alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XWalletAddress'
        - $ref: '#/components/parameters/XWalletSignature'
        - $ref: '#/components/parameters/XWalletNonce'
      responses:
        '200':
          description: Alert acknowledged
  /config/alerts/{id}/mute:
    post:
      summary: Mute an alert for a duration in minutes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XWalletAddress'
        - $ref: '#/components/parameters/XWalletSignature'
        - $ref: '#/components/parameters/XWalletNonce'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                minutes:
                  type: integer
                reason:
                  type: string
      responses:
        '200':
          description: Alert muted
  /config/alerts/{id}/unmute:
    post:
      summary: Remove alert mute state
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XWalletAddress'
        - $ref: '#/components/parameters/XWalletSignature'
        - $ref: '#/components/parameters/XWalletNonce'
      responses:
        '200':
          description: Alert unmuted
...
components:
  parameters:
...
    XWalletAddress:
      in: header
      name: x-wallet-address
      required: true
      schema:
        type: string
    XWalletSignature:
      in: header
      name: x-wallet-signature
      required: true
      schema:
        type: string
    XWalletNonce:
      in: header
      name: x-wallet-nonce
      required: true
      schema:
        type: string